feat(formulus): move Settings from drawer dropdown to Settings screen
